Do not download kind binaries if already there

+++ kind/k8s-create.sh
@@ -5,21 +5,28 @@ set -x
 
 KUBECTL_BIN="/usr/local/bin/kubectl"
 KIND_BIN="/usr/local/bin/kind"
+CLUSTER_NAME="kind"
 
-curl -Lo /tmp/kind https://github.com/kubernetes-sigs/kind/releases/download/v0.3.0/kind-linux-amd64
-chmod +x /tmp/kind
-sudo mv /tmp/kind "$KIND_BIN"
+. "$DIR/env.sh"
+
+if [ ! -e $KIND_BIN ]; then
+    curl -Lo /tmp/kind https://github.com/kubernetes-sigs/kind/releases/download/v0.3.0/kind-linux-amd64
+    chmod +x /tmp/kind
+    sudo mv /tmp/kind "$KIND_BIN"
+fi
 
 # Download kubectl, which is a requirement for using kind.
-K8S_VERSION="v1.13.0"
-curl -Lo /tmp/kubectl https://storage.googleapis.com/kubernetes-release/release/"$K8S_VERSION"/bin/linux/amd64/kubectl
-chmod +x /tmp/kubectl
-sudo mv /tmp/kubectl "$KUBECTL_BIN"
+if [ ! -e $KUBECTL_BIN ]; then
+    K8S_VERSION="v1.13.0"
+    curl -Lo /tmp/kubectl https://storage.googleapis.com/kubernetes-release/release/"$K8S_VERSION"/bin/linux/amd64/kubectl
+    chmod +x /tmp/kubectl
+    sudo mv /tmp/kubectl "$KUBECTL_BIN"
+fi
 
 DIR=$(cd "$(dirname "$0")"; pwd -P)
-kind create cluster --config "$DIR/kind-config.yaml"
+kind create cluster --config "$DIR/kind-config.yaml" --name "$CLUSTER_NAME" 
 
-export KUBECONFIG=$(kind get kubeconfig-path)
+export KUBECONFIG=$(kind get kubeconfig-path --name="$CLUSTER_NAME")
 
 # this for loop waits until kubectl can access the api server that kind has created
 JSONPATH='{range .items[*]}{@.metadata.name}:{range @.status.conditions[*]}{@.type}={@.status};{end}{end}'
